test: refactor transaction propagation tests to use MikroORM.init and describe.each pattern
